Refactor modules/packages: "foreign" to "ffm"
FFM seems to be what the Java platform is now calling
Panama now that it is GA, so let's rename now to match that
nomenclature.
