+++ secp256k1-examples-java/build.gradle
@@ -12,7 +12,7 @@ ext.moduleName = 'org.bitcoinj.secp256k1.examples'
 dependencies {
     implementation project(':secp256k1-api')
     runtimeOnly project(':secp256k1-bouncy')
-    runtimeOnly project(':secp256k1-foreign')
+    runtimeOnly project(':secp256k1-ffm')
 }
 
 jar {
