+++ secp256k1-ffm/build.gradle
@@ -6,20 +6,16 @@ tasks.withType(JavaCompile).configureEach {
     //options.release = 22
 }
 
-ext.moduleName = 'org.bitcoinj.secp256k1.bitcoinj'
+ext.moduleName = 'org.bitcoinj.secp256k1.ffm'
 
 dependencies {
     api project(':secp256k1-api')
-    api("org.bitcoinj:bitcoinj-core:0.17-alpha3");
-
-    implementation project(':secp256k1-bouncy')
-    implementation project(':secp256k1-ffm')
 }
 
 jar {
     inputs.property("moduleName", moduleName)
     manifest {
-        attributes  'Implementation-Title': 'Secp256k1 bitcoinj support',
+        attributes  'Implementation-Title': 'Secp256k1 libsecp256k1 Foreign Function & Memory implementation',
                 'Implementation-Version': archiveVersion.get()
     }
 }
