+++ secp256k1-ffm/src/main/resources/META-INF/services/org.bitcoinj.secp256k1.api.Secp256k1Provider
@@ -1,0 +1,1 @@
+org.bitcoinj.secp256k1.ffm.ForeignProvider
