--- secp256k1-foreign/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-tasks.withType(JavaCompile).configureEach {
-    //options.release = 22
-}
-
-ext.moduleName = 'org.bitcoinj.secp256k1.foreign'
-
-dependencies {
-    api project(':secp256k1-api')
-}
-
-jar {
-    inputs.property("moduleName", moduleName)
-    manifest {
-        attributes  'Implementation-Title': 'Secp256k1 libsecp256k1 Foreign Function & Memory implementation',
-                'Implementation-Version': archiveVersion.get()
-    }
-}
