--- secp256k1-foreign/src/main/java/org/bitcoinj/secp256k1/foreign/jextract/secp256k1_pubkey.java
@@ -1,175 +1,0 @@
-/*
- * Copyright 2023-2024 secp256k1-jdk Developers.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-// Generated by jextract
-
-package org.bitcoinj.secp256k1.foreign.jextract;
-
-import java.lang.invoke.*;
-import java.lang.foreign.*;
-import java.nio.ByteOrder;
-import java.util.*;
-import java.util.function.*;
-import java.util.stream.*;
-
-import static java.lang.foreign.ValueLayout.*;
-import static java.lang.foreign.MemoryLayout.PathElement.*;
-
-/**
- * {@snippet lang=c :
- * struct {
- *     unsigned char data[64];
- * }
- * }
- */
-public class secp256k1_pubkey {
-
-    secp256k1_pubkey() {
-        // Should not be called directly
-    }
-
-    private static final GroupLayout $LAYOUT = MemoryLayout.structLayout(
-        MemoryLayout.sequenceLayout(64, secp256k1_h.C_CHAR).withName("data")
-    ).withName("$anon$74:9");
-
-    /**
-     * The layout of this struct
-     */
-    public static final GroupLayout layout() {
-        return $LAYOUT;
-    }
-
-    private static final SequenceLayout data$LAYOUT = (SequenceLayout)$LAYOUT.select(groupElement("data"));
-
-    /**
-     * Layout for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static final SequenceLayout data$layout() {
-        return data$LAYOUT;
-    }
-
-    private static final long data$OFFSET = 0;
-
-    /**
-     * Offset for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static final long data$offset() {
-        return data$OFFSET;
-    }
-
-    /**
-     * Getter for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static MemorySegment data(MemorySegment struct) {
-        return struct.asSlice(data$OFFSET, data$LAYOUT.byteSize());
-    }
-
-    /**
-     * Setter for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static void data(MemorySegment struct, MemorySegment fieldValue) {
-        MemorySegment.copy(fieldValue, 0L, struct, data$OFFSET, data$LAYOUT.byteSize());
-    }
-
-    private static long[] data$DIMS = { 64 };
-
-    /**
-     * Dimensions for array field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static long[] data$dimensions() {
-        return data$DIMS;
-    }
-    private static final VarHandle data$ELEM_HANDLE = data$LAYOUT.varHandle(sequenceElement());
-
-    /**
-     * Indexed getter for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static byte data(MemorySegment struct, long index0) {
-        return (byte)data$ELEM_HANDLE.get(struct, 0L, index0);
-    }
-
-    /**
-     * Indexed setter for field:
-     * {@snippet lang=c :
-     * unsigned char data[64]
-     * }
-     */
-    public static void data(MemorySegment struct, long index0, byte fieldValue) {
-        data$ELEM_HANDLE.set(struct, 0L, index0, fieldValue);
-    }
-
-    /**
-     * Obtains a slice of {@code arrayParam} which selects the array element at {@code index}.
-     * The returned segment has address {@code arrayParam.address() + index * layout().byteSize()}
-     */
-    public static MemorySegment asSlice(MemorySegment array, long index) {
-        return array.asSlice(layout().byteSize() * index);
-    }
-
-    /**
-     * The size (in bytes) of this struct
-     */
-    public static long sizeof() { return layout().byteSize(); }
-
-    /**
-     * Allocate a segment of size {@code layout().byteSize()} using {@code allocator}
-     */
-    public static MemorySegment allocate(SegmentAllocator allocator) {
-        return allocator.allocate(layout());
-    }
-
-    /**
-     * Allocate an array of size {@code elementCount} using {@code allocator}.
-     * The returned segment has size {@code elementCount * layout().byteSize()}.
-     */
-    public static MemorySegment allocateArray(long elementCount, SegmentAllocator allocator) {
-        return allocator.allocate(MemoryLayout.sequenceLayout(elementCount, layout()));
-    }
-
-    /**
-     * Reinterprets {@code addr} using target {@code arena} and {@code cleanupAction} (if any).
-     * The returned segment has size {@code layout().byteSize()}
-     */
-    public static MemorySegment reinterpret(MemorySegment addr, Arena arena, Consumer<MemorySegment> cleanup) {
-        return reinterpret(addr, 1, arena, cleanup);
-    }
-
-    /**
-     * Reinterprets {@code addr} using target {@code arena} and {@code cleanupAction} (if any).
-     * The returned segment has size {@code elementCount * layout().byteSize()}
-     */
-    public static MemorySegment reinterpret(MemorySegment addr, long elementCount, Arena arena, Consumer<MemorySegment> cleanup) {
-        return addr.reinterpret(layout().byteSize() * elementCount, arena, cleanup);
-    }
-}
-
