--- secp256k1-foreign/src/main/resources/META-INF/services/org.bitcoinj.secp256k1.api.Secp256k1Provider
@@ -1,1 +1,0 @@
-org.bitcoinj.secp256k1.foreign.ForeignProvider
