+++ secp256k1-integration-test/build.gradle
@@ -11,7 +11,7 @@ ext.moduleName = 'org.bitcoinj.secp256k1.integration'
 dependencies {
     implementation project(':secp256k1-api')
     implementation project(':secp256k1-bouncy')
-    implementation project(':secp256k1-foreign')
+    implementation project(':secp256k1-ffm')
     implementation project(':secp256k1-sandbox')
 }
