+++ secp256k1-integration-test/src/test/java/org/bitcoinj/secp256k1/integration/CurveTest.java
@@ -20,7 +20,7 @@ import org.bitcoinj.secp256k1.api.P256k1PubKey;
 import org.bitcoinj.secp256k1.api.Secp256k1;
 import org.bitcoinj.secp256k1.bouncy.Bouncy256k1;
 import org.bitcoinj.secp256k1.bouncy.BouncyPrivKey;
-import org.bitcoinj.secp256k1.foreign.Secp256k1Foreign;
+import org.bitcoinj.secp256k1.ffm.Secp256k1Foreign;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigInteger;
