+++ settings.gradle
@@ -2,7 +2,7 @@ rootProject.name = 'secp256k1-jdk'
 
 include 'secp256k1-api'               // API interface library
 include 'secp256k1-bouncy'            // Bouncy Castle implementation
-include 'secp256k1-foreign'           // Java Foreign Memory & Function ("Panama") implementation
+include 'secp256k1-ffm'               // Java Foreign Memory & Function ("Panama") implementation
 include 'secp256k1-bitcoinj'          // bitcoinj integration utilities and tests (P2TR address generation, etc.)
 include 'secp256k1-integration-test'  // Integration tests of API implementations (foreign and bouncy)
 include 'secp256k1-examples-java'     // Java examples
